imporved ui and added productivity and streak features

+++ src/main/ipcHandlers.js
@@ -4,12 +4,27 @@ import * as usageService from './services/usageService';
 import * as userService from './services/userService';
 import * as exportService from './services/exportService';
 import { setCurrentActiveGoal } from './services/trackingService'
+const NodeCache = require('node-cache');
+const myCache = new NodeCache();
 
 
 function registerIpcHandlers() {
     //goal handlers
     ipcMain.handle('get-goals',async () => {
   return Promise.resolve(goalService.getGoals());});
+  ipcMain.handle('get-today-productive-time', () => goalService.getTodayProductiveTime());
+  ipcMain.handle('get-global-log-streak',(event,minTime) => {
+    const todayDate = new Date().toLocaleDateString('en-CA', {
+  timeZone: 'Asia/Kolkata'
+});
+const cacheKey = `streak-${todayDate}`;
+let streak = myCache.get(cacheKey);
+if (streak == null) {
+    streak = goalService.getGlobalProductiveStreak(minTime) ; // Your streak calc function
+    myCache.set(cacheKey, streak, 60*60*24); // Cache for 24hr
+  }
+  return streak;
+  });
     ipcMain.on('add-goal', (event,goal) => goalService.addGoal(goal));
     ipcMain.on('delete-goal', (event,goalId) => goalService.deleteGoalWithSubtasks(goalId));
     ipcMain.on('update-goal',(event,goal) => goalService.update_goal(goal));
